requirements.txt
flask
line-bot-sdk
openai
streamlit
pandas
gunicorn
gspread
google-auth
